requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.6.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.26.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
scikit-learn>=1.3.0
torch>=2.1.0
transformers>=4.36.0
tiktoken>=0.5.0
python-multipart>=0.0.6
aiofiles>=23.2.0
